+++ lib/src/enums/post_message_type.dart
@@ -8,4 +8,5 @@ enum PostMessageType {
   copyToClipboard,
   getImage,
   custom,
+  didChangeAppLifecycleState,
 }